+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.5.2 (Next)
 ============
 
+* [#215](https://github.com/ruby-grape/grape-entity/pull/217): `#delegate_attribute` should not delegate to methods included with `Kernel`. - [@maltoe](https://github.com/maltoe)
 * Your contribution here.
 
 0.5.1 (2016-4-4)